backend/requirements.txt
flask==3.1.2
flask-cors==5.0.1
tensorflow>=2.15.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
gunicorn>=21.2.0
